Refactor/solve sonarqube warnings

+++ src/test/java/com/example/fairshareapp/controller/GastoControllerTest.java
@@ -16,6 +16,7 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Collections;
 import java.util.List;
@@ -64,7 +65,7 @@ class GastoControllerTest {
         GastoDetalleDTO detalle = GastoDetalleDTO.builder()
                 .id(1L)
                 .descripcion("Cena")
-                .monto(100.0)
+                .monto(BigDecimal.valueOf(100.0))
                 .espacioId(10L)
                 .regla(ReglaDivision.EQUITATIVA)
                 .participantes(Collections.emptyList())
@@ -99,7 +100,7 @@ class GastoControllerTest {
         GastoDetalleDTO detalle = GastoDetalleDTO.builder()
                 .id(1L)
                 .descripcion("Lote 1")
-                .monto(250.0)
+                .monto(BigDecimal.valueOf(250.0))
                 .build();
 
         when(gastoService.registrarLoteGastos(eq(10L), any())).thenReturn(List.of(detalle));
@@ -131,7 +132,7 @@ class GastoControllerTest {
         GastoDetalleDTO detalle = GastoDetalleDTO.builder()
                 .id(5L)
                 .descripcion("Gasto recuperado")
-                .monto(300.0)
+                .monto(BigDecimal.valueOf(300.0))
                 .fecha(LocalDate.of(2026, 9, 7))
                 .build();
